+++ app/src/main/java/com/example/drivingefficiencyapp/MainMenuActivity.kt
@@ -12,10 +12,11 @@ package com.example.drivingefficiencyapp
 
 import android.content.Intent
 import android.os.Bundle
-import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
+import com.example.drivingefficiencyapp.databinding.MainMenuActivityBinding
 
 class MainMenuActivity : AppCompatActivity() {
+    private lateinit var binding: MainMenuActivityBinding // View Binding
     /**
      * Creates the main menu screen and sets up the buttons to start a drive or view past trips.
      *
@@ -25,19 +26,18 @@ class MainMenuActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState) //Re-create the saved state
         supportActionBar?.hide() //Hid the action bar which references the app name
-        setContentView(R.layout.main_menu_activity) //Set the layout using main_menu_activity.xml layout
 
-        //Set up the buttons to start a drive or view past trips, using the IDs from main_menu_activity.xml
-        val startDriveButton = findViewById<Button>(R.id.startDriveButton)
-        val viewTripsButton = findViewById<Button>(R.id.viewTripsButton)
+        binding = MainMenuActivityBinding.inflate(layoutInflater)
+
+        setContentView(binding.root) //Set the layout using main_menu_activity.xml layout
 
         //Set up the button listeners on the buttons to start the StartDriveActivity or TripsActivity
-        startDriveButton.setOnClickListener {
+        binding.startDriveButton.setOnClickListener {
             val intent = Intent(this, StartDriveActivity::class.java)
             startActivity(intent)
         }
 
-        viewTripsButton.setOnClickListener {
+        binding.viewTripsButton.setOnClickListener {
             val intent = Intent(this, TripsActivity::class.java)
             startActivity(intent)
         }